+++ src/chanjo2/meta/handle_completeness_tasks.py
@@ -3,7 +3,7 @@ import subprocess
 from multiprocessing import Manager, Pool
 from typing import Dict, List, Tuple
 
-LOG = logging.getLogger("uvicorn.access")
+LOG = logging.getLogger(__name__)
 INTERVAL_CHUNKS = 50
 CHROM_INDEX = 0
 START_INDEX = 1