+++ src/chanjo2/__init__.py
@@ -1,4 +1,4 @@
 from dotenv import load_dotenv
 
-__version__ = "1.8"
+__version__ = "1.9"
 load_dotenv()