+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Updated PR template
 - Generalised issue templates to make them more user-friendly for people outside our organisation
 - Moved logging setup out of app lifespan and db initialisation logic
+- Switch to __main__ logger
 ### Fixed
 - Updated version of external images used in GitHub actions