+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 - Updated PR template
 - Generalised issue templates to make them more user-friendly for people outside our organisation
 - Moved logging setup out of app lifespan and db initialisation logic
-- Switch to __main__ logger
+- Switch to __main__ logger and removed unused logger from multiprocessing code
 ### Fixed
 - Updated version of external images used in GitHub actions