+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [unreleased]
 ### Added
 - Condensed `/coverage/d4/genes/summary` for condensed stats over a gene list
+### Changed
+- GitHub tests action to use d4tools 0.3.10
 ### Fixed
 - Updated dependencies to address dependabot's security alerts
 - Use a base image containing d4tools 0.3.10 in Dockerfile